eliminated box/unbox

+++ src/main/java/org/eclipselabs/garbagecat/util/jdk/JdkMath.java
@@ -175,9 +175,8 @@ public final class JdkMath {
      */
     public static int calcThroughput(final int currentDuration, final long currentTimestamp, final int priorDuration,
             final long priorTimestamp) {
-        long timeTotal = currentTimestamp + Long.valueOf(currentDuration).longValue() - priorTimestamp;
-        long timeNotGc = timeTotal - Long.valueOf(currentDuration).longValue()
-                - Long.valueOf(priorDuration).longValue();
+        long timeTotal = currentTimestamp + currentDuration - priorTimestamp;
+		long timeNotGc = timeTotal - currentDuration - priorDuration;
         BigDecimal throughput = new BigDecimal(timeNotGc);
         throughput = throughput.divide(new BigDecimal(timeTotal), 2, RoundingMode.HALF_EVEN);
         throughput = throughput.movePointRight(2);
